refactor: change directory for CreateNoteForm and DialogCreateNote components
